safari-534.53-branch/LayoutTests/svg/dynamic-updates/SVGFEPointLightElement-svgdom-z-prop-expected.txt

CONSOLE MESSAGE: line 51: TypeError: 'undefined' is not an object (evaluating 'pointLight.z.baseVal = 30')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'z' property of the SVGFEPointLightElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL pointLight.z.baseVal should be 100. Threw exception TypeError: 'undefined' is not an object (evaluating 'pointLight.z.baseVal')
